Created PublicDefaultTemplate component and moved template to it Created Signup component - moved sign up logic Updated Register to use above defined components
