+++ nt.webclient/vuejs/nt/src/pages/public/Register.vue
@@ -1,163 +1,17 @@
 <template>
-  <div class="row">
-    <div class="col-lg-9">More details will come here - UNDER CONSTRUCTION</div>
-    <div class="col-lg-3">
-      <div class="card card-block rounded shadow shadow-sm">
-        <div class="card-header bg-primary text-light text-uppercase">
-          <div class="card-title align-middle">
-            <h5 class="mb-0">Sign Up</h5>
-          </div>
-        </div>
-        <div class="card-body">
-          <form class="form needs-validation" @submit="onSubmit">
-            <div class="form-group">
-              <input
-                type="text"
-                v-model="userName"
-                v-bind:class="
-                  hasError('userName')
-                    ? 'form-control block is-invalid'
-                    : 'form-control block'
-                "
-                placeholder="Username"
-              />
-            </div>
-            <div
-              v-bind:class="
-                hasError('userName') ? 'text-danger text-left' : 'd-none'
-              "
-            >
-              <small>Username cannot be empty</small>
-            </div>
-            <div class="form-group">
-              <input
-                type="password"
-                v-model="password"
-                v-bind:class="
-                  hasError('password')
-                    ? 'form-control block is-invalid'
-                    : 'form-control block'
-                "
-                placeholder="Password"
-              />
-            </div>
-            <div
-              v-bind:class="
-                hasError('password') ? 'text-danger text-left' : 'd-none'
-              "
-            >
-              <small>Password cannot be empty</small>
-            </div>
-            <div class="form-group">
-              <input
-                type="password"
-                v-model="confirmPassword"
-                v-bind:class="
-                  hasError('confirmPassword')
-                    ? 'form-control block is-invalid'
-                    : 'form-control block'
-                "
-                placeholder="Confirm Password"
-              />
-            </div>
-            <div
-              v-bind:class="
-                hasError('confirmPassword') ? 'text-danger text-left' : 'd-none'
-              "
-            >
-              <small>Password does not match</small>
-            </div>
-
-            <div class="form-group">
-              <input type="submit" class="btn btn-block btn-primary" value="Submit" />
-            </div>
-            <div v-bind:class="showServerMessage()">
-              <small>{{serverMessage}}</small>
-            </div>
-          </form>
-          <div>
-            Not a member ?
-            <router-link to="/register">Sign up here</router-link>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
+  <PublicDefaultTemplate>
+    <SignUp />
+  </PublicDefaultTemplate>
 </template>
 
 <script>
-import axios from "axios";
+import SignUp from "../../components/public/Signup";
+import PublicDefaultTemplate from "../../components/public/PublicDefaultTemplate";
 export default {
   name: "Register",
-  data() {
-    return {
-      errors: [],
-      userName: "",
-      password: "",
-      confirmPassword: "",
-      serverMessage: "",
-      hasServerError: false,
-    };
-  },
-  methods: {
-    hasError(key) {
-      return this.errors.indexOf(key) != -1;
-    },
-    showServerMessage() {
-      if (!this.serverMessage) {
-        return "d-none";
-      }
-
-      return this.hasServerError
-        ? "text-danger text-left"
-        : "text-success text-left";
-    },
-    async onSubmit(e) {
-      e.preventDefault();
-
-      if (!this.validateForm()) {
-        // has Error
-        console.log(this.userName);
-        return;
-      }
-
-      const userDetails = {
-        userName: this.userName,
-        passKey: this.password,
-        displayName: this.userName,
-      };
-      const response = await axios.post(
-        "https://localhost:44353/api/User/CreateUser",
-        userDetails
-      );
-
-      this.hasServerError = response.data.hasError;
-      if (this.hasServerError) {
-        this.serverMessage = response.data.errorMessage;
-      } else {
-        this.serverMessage = "User created successfully.";
-      }
-      console.log(this.showServerMessage());
-    },
-    validateForm() {
-      let isValidFlag = true;
-      this.errors = [];
-      if (!this.userName) {
-        this.errors.push("userName");
-        isValidFlag = false;
-      }
-
-      if (!this.password) {
-        this.errors.push("password");
-        isValidFlag = false;
-      }
-
-      if (this.password != this.confirmPassword) {
-        this.errors.push("confirmPassword");
-        isValidFlag = false;
-      }
-      return isValidFlag;
-    },
+  components: {
+    SignUp,
+    PublicDefaultTemplate,
   },
 };
 </script>
